Move TextUtils to utils package.
