+++ geoclient-parser/src/main/java/gov/nyc/doitt/gis/geoclient/parser/Input.java
@@ -15,8 +15,8 @@
  */
 package gov.nyc.doitt.gis.geoclient.parser;
 
-import gov.nyc.doitt.gis.geoclient.parser.token.TextUtils;
 import gov.nyc.doitt.gis.geoclient.parser.util.Assert;
+import gov.nyc.doitt.gis.geoclient.parser.util.TextUtils;
 
 
 /**
