+++ geoclient-parser/src/main/java/gov/nyc/doitt/gis/geoclient/parser/util/TextUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package gov.nyc.doitt.gis.geoclient.parser.token;
+package gov.nyc.doitt.gis.geoclient.parser.util;
 
 import org.apache.commons.lang3.RegExUtils;
 import org.apache.commons.lang3.StringUtils;
